Spin IDE 1.0.0.11 -----------------
Changes in this release
-----------------------

1. Added flag so iTunes may be used to copy files in and out of the app sandbox.

2. Projects can be emailed using the Share button.

3. Projects can be imported from email. Select the exported pzip file in the iPad mail client, tap share, and select Spin from the list of export possibilities.

4. Spin files can be imported from email. Select the spin file, then follow the directions for a project.

+++ SpinIDE/Terminal/TerminalView.m
@@ -611,11 +611,11 @@ static TerminalView *this;
                 theText = [theText substringFromIndex: 1];
             } else {
                 if (index < 0)
-                    index = theText.length;
+                    index = (int) theText.length;
                 NSString *text = [theText substringToIndex: index];
                 NSRange range = terminalOutputView.selectedRange;
                 int length = (int) text.length;
-                int pos = range.location;
+                int pos = (int) range.location;
                 while (length-- && pos < terminalOutputView.text.length && [terminalOutputView.text characterAtIndex: pos] != '\n') {
                     ++range.length;
                     ++pos;
